fix(swingset): factor out getDecidedPromises()
cleanupAfterTerminatedVat() had a loop which identified all the
promises (kpids) which were decided by a vat, so it could reject all
of them as the vat was being terminated. We will have a need for this
list of kpids during vat upgrade too, so refactor it out into a
separate kernelKeeper method, and change terminateVat() to fetch the
list before calling cleanupAfterTerminatedVat().